start of making lots of small files rather than a few big ones
