+++ SetupHistos_only.hh
@@ -1,0 +1,12 @@
+#ifndef SETUPHISTOS_ONLY_H
+#define SETUPHISTOS_ONLY_H
+
+#include <TH3.h>
+#include <TH2.h>
+#include <TH1.h>
+#include <TList.h>
+#include <cstdio>
+
+void SetupHistos(TList *outlist);
+
+#endif
